refactor group_structure
